+++ src/main/java/com/simibubi/create/api/packager/unpacking/UnpackingHandler.java
@@ -1,4 +1,4 @@
-package com.simibubi.create.api.unpacking;
+package com.simibubi.create.api.packager.unpacking;
 
 import java.util.List;
