+++ src/main/java/com/simibubi/create/impl/unpacking/DefaultUnpackingHandler.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.jetbrains.annotations.Nullable;
 
-import com.simibubi.create.api.unpacking.UnpackingHandler;
+import com.simibubi.create.api.packager.unpacking.UnpackingHandler;
 import com.simibubi.create.content.logistics.stockTicker.PackageOrder;
 
 import net.minecraft.core.BlockPos;
